style: 열거형 Message를 구조체JuiceMaker의 중첩타입으로 이동
